
\name{TASTE.S}
\alias{TASTE.S}
\title{Extracting the proteins (Step3)}
\usage{
TASTE.S(pro_name, jefs, lower, upper, lambda.v, gamma)
}
\arguments{
\item{lower}{lower limit of the l-0 norm of the vector v a sparse loading vector, corresponding to columns of joint matrix.}

\item{upper}{upper limit of the l-0 norm of the vector v a sparse loading vector, corresponding to columns of joint matrix.}


\item{gamma}{step-size.}
}
\description{
Extracting the proteins (Step3)
}
